adding updates to seed route, and adding .env

+++ server/server.js
@@ -1,3 +1,4 @@
+require('dotenv').config();
 const express = require('express');
 const { ApolloServer } = require('apollo-server-express');
 const path = require('path');
@@ -31,20 +32,23 @@ app.get('/', (req, res) => {
 const startApolloServer = async (typeDefs, resolvers) => {
   await server.start();
   server.applyMiddleware({ app });
-  
+
   db.once('open', () => {
-    app.get('/seedDatabase', async (req, res)=> {
-      await Tech.deleteMany({});
+    app.post('/seedDatabase', async (req, res) => {
+      SEEDPASS=DB_PASSWORD;
+      if(req.body.SEEDPASS === process.env.SEEDPASS){ await Tech.deleteMany({});
       const technologies = await Tech.insertMany(techData);
       res.json(technologies)
-    })
+      }
+    });
+
     app.listen(PORT, () => {
       console.log(`API server running on port ${PORT}!`);
       console.log(`Use GraphQL at http://localhost:${PORT}${server.graphqlPath}`);
     })
   })
-  };
-  
+};
+
 // Call the async function to start the server
-  startApolloServer(typeDefs, resolvers);
- 
+startApolloServer(typeDefs, resolvers);
+
